pyproject.toml
[project]
name = "voice-input"
version = "0.1.0"
description = "iPhone voice input to Windows typing via OpenAI transcription"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "openai",
  "pydantic",
  "pyautogui",
  "python-dotenv",
  "python-multipart",
  "requests",
  "uvicorn[standard]",
]
